Initial Database setup and default login page
Not sure exactly what we will need in database so far, feel free to add anything deemed necessary. I put in the automatically made login page which is almost 300 lines and seems very complicated so we might want to see how we are going to use that/if we want to simplify it.